manual: Rename nodes to silence Emacs name check
* doc/org-manual.org (Advanced Export Configuration):
(Extracting Source Code): Rename "Hooks" subheading so that ox-texinfo
does not make it unique by appending " (N)".

The Emacs repo has a check for problematic characters in info nodes
that flagged "Hooks (2)" (bug#52126).  As mentioned in that thread,
this case is actually probably not problematic because the name
doesn't _start_ with "(".  However, it's easy enough to sidestep the
warning, and adding the extra context doesn't hurt.